Group analysis as one test
